Move base component to core
